+++ korma/src/commonMain/kotlin/korlibs/math/geom/bezier/Curve.kt
@@ -6,14 +6,14 @@ import korlibs.math.interpolation.*
 interface Curve {
     val order: Int
     fun getBounds(): Rectangle
-    fun normal(t: Double): Point
-    fun tangent(t: Double): Point
-    fun calc(t: Double): Point
-    fun ratioFromLength(length: Double): Double = TODO()
-    val length: Double
+    fun normal(t: Float): Point
+    fun tangent(t: Float): Point
+    fun calc(t: Float): Point
+    fun ratioFromLength(length: Float): Float = TODO()
+    val length: Float
     // @TODO: We should probably have a function to get ratios in the function to place the points maybe based on inflection points?
     fun recommendedDivisions(): Int = DEFAULT_STEPS
-    fun calcOffset(t: Double, offset: Double): Point = calc(t) + normal(t) * offset
+    fun calcOffset(t: Float, offset: Float): Point = calc(t) + normal(t) * offset
 
 
     companion object {
@@ -25,7 +25,7 @@ interface Curve {
 internal fun Curve._getPoints(count: Int = this.recommendedDivisions(), equidistant: Boolean = false, out: PointArrayList = PointArrayList()): PointList {
     val curveLength = length
     Ratio.forEachRatio(count) { ratio ->
-        val t = if (equidistant) ratioFromLength(ratio.toDouble() * curveLength) else ratio.toDouble()
+        val t = if (equidistant) ratioFromLength(ratio.toFloat() * curveLength) else ratio.toFloat()
         //println("${this::class.simpleName}: ratio: $ratio, point=$point")
         out.add(calc(t))
     }
